@@ -23,78 +23,74 @@
    the patch version to avoid breakages due to api changes
    which usually occur on minor and major version numbers.
 */
-
-repositories {
-    mavenCentral()
-    mavenLocal()
-
-    flatDir {
-        dirs 'libs'
-    }
-    maven { url "https://jitpack.io" }
+plugins {
+    id 'com.android.application'
 }
-
-
-dependencies {
-    //    compile project(':control_app_lib')
-    compile fileTree(include: ['*.jar'], dir: 'libs')
-    ///////////////
-    compile 'org.ros.android_core:android_10:[0.2,0.3)'
-    compile 'org.ros.android_core:android_15:[0.2,0.3)'
-    compile 'com.github.rosjava.android_extras:gingerbread:[0.2,0.3)'
-    compile 'org.ros.rosjava_messages:tf2_msgs:[0.5,0.6)'
-    compile 'com.google.code.gson:gson:2.5'
-    compile 'com.android.support:support-v13:23.2.0'
-    compile 'com.android.support:appcompat-v7:23.2.0'
-    compile 'com.android.support:cardview-v7:23.2.0'
-    compile 'com.android.support:recyclerview-v7:23.2.0'
-    compile 'org.osmdroid:osmdroid-android:5.2@aar'
-    compile(name:'osmbonuspack_v5.8.1', ext:'aar')
-    compile 'org.apache.commons:commons-lang3:3.3.2'
-    compile 'com.squareup.okhttp:okhttp:2.6.0'
-    compile 'com.android.support:support-v4:23.2.0'
-    compile 'com.github.amlcurran.showcaseview:library:5.4.1'
-    compile project(':control_app_lib')
-}
-
-apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 23
+    compileSdk 32
     defaultConfig {
-        minSdkVersion 15
-        targetSdkVersion 23
+        minSdk 21
+        targetSdk 32
         multiDexEnabled true
         versionCode 1
         versionName "1.0"
 
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
     }
-    sourceSets {
-        main { java.srcDirs = ['src/main/java'] }
-        debug { java.srcDirs = ['src/debug/java'] }
-    }
+//    sourceSets {
+//        main { java.srcDirs = ['src/main/java'] }
+//        debug { java.srcDirs = ['src/debug/java'] }
+//    }
 }
 
 dependencies {
-    // Required -- JUnit 4 framework
-    testCompile 'junit:junit:4.12'
-    // Optional -- Mockito framework
-    testCompile 'org.mockito:mockito-core:1.10.19'
+    //    implementation project(':control_app_lib')
+    implementation fileTree(include: ['*.jar'], dir: 'libs')
+    ///////////////
+    implementation 'org.ros.android_core:android_10:[0.2,0.3)'
+    implementation 'org.ros.android_core:android_15:[0.2,0.3)'
+    implementation 'com.github.rosjava.android_extras:gingerbread:[0.2,0.3)'
+    implementation 'org.ros.rosjava_messages:tf2_msgs:[0.5,0.6)'
+    implementation 'com.google.code.gson:gson:2.8.6'
+    implementation 'androidx.legacy:legacy-support-v13:1.0.0'
+    implementation 'androidx.appcompat:appcompat:1.5.0'
+    implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'androidx.recyclerview:recyclerview:1.2.1'
+    implementation 'org.osmdroid:osmdroid-android:5.2@aar'
+    implementation(name: 'osmbonuspack_v5.8.1', ext: 'aar')
+    implementation 'org.apache.commons:commons-lang3:3.3.2'
+    implementation 'com.squareup.okhttp:okhttp:2.6.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'com.github.amlcurran.showcaseview:library:5.4.1'
+    implementation project(':control_app_lib')
 
-    androidTestCompile 'com.android.support:support-annotations:23.2.0'
-    androidTestCompile 'com.android.support.test:runner:0.4'
-    androidTestCompile 'com.android.support.test:rules:0.4'
+    implementation 'org.ros.rosjava_bootstrap:message_generation:0.3.3'
+    implementation 'org.ros.rosjava_messages:nav_msgs:1.12.7'
+    implementation 'org.ros.rosjava_messages:geometry_msgs:1.12.7'
+
+    // Required -- JUnit 4 framework
+    testImplementation 'junit:junit:4.13.2'
+    // Optional -- Mockito framework
+    testImplementation 'org.mockito:mockito-core:1.10.19'
+
+    androidTestImplementation 'androidx.annotation:annotation:1.4.0'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.3'
+    androidTestImplementation 'androidx.test:rules:1.4.0'
     // Optional -- Hamcrest library
-    androidTestCompile 'org.hamcrest:hamcrest-library:1.3'
+    androidTestImplementation 'org.hamcrest:hamcrest-library:1.3'
     // Optional -- UI testing with Espresso
-    androidTestCompile 'com.android.support.test.espresso:espresso-core:2.2.1'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
     // Optional -- UI testing with UI Automator
-//    androidTestCompile 'com.android.support.test.uiautomator:uiautomator-v15'
+//    androidTestImplementation 'com.android.support.test.uiautomator:uiautomator-v15'
 }
 
 // Needed to resolve app vs test dependencies, specifically, transitive dependencies of
 // libraryq and libraryz. Forcing the use of the smaller version after regression testing.
 configurations.all {
     resolutionStrategy.force 'com.google.code.findbugs:jsr305:1.3.9'
+}
+
+configurations {
+    all*.exclude group: 'com.google.guava', module: 'listenablefuture'
 }
