@@ -36,11 +36,11 @@ dependencies {
 apply plugin: 'com.android.library'
 
 android {
-  compileSdkVersion 23
-  defaultConfig {
-    minSdkVersion 15
-    targetSdkVersion 23
-    versionCode 1
-    versionName "1.0"
-  }
+    compileSdkVersion 32
+    defaultConfig {
+        minSdkVersion 21
+        targetSdkVersion 32
+        versionCode 1
+        versionName "1.0"
+    }
 }
