@@ -2,7 +2,7 @@ package com.robotca.ControlApp.Core;
 
 import android.os.Bundle;
 import android.preference.PreferenceManager;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.util.Log;
 
 import com.robotca.ControlApp.ControlApp;
