@@ -5,9 +5,9 @@ import android.app.Activity;
 import android.app.Dialog;
 import android.content.DialogInterface;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.v4.app.DialogFragment;
-import android.support.v7.app.AlertDialog;
+import androidx.annotation.NonNull;
+import androidx.fragment.app.DialogFragment;
+import androidx.appcompat.app.AlertDialog;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.widget.CheckBox;
@@ -16,8 +16,6 @@ import android.widget.Toast;
 
 import com.robotca.ControlApp.Core.RobotInfo;
 import com.robotca.ControlApp.R;
-
-import java.util.UUID;
 
 /**
  * Dialog for adding or editing a Robot.
