@@ -6,7 +6,7 @@ package com.robotca.ControlApp.Fragments;
  * Created by kennethspear on 3/28/16.
  */
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
